requirements.txt
robotframework
robotframework-seleniumlibrary
robotframework-requests
robotframework-databaselibrary
robotframework-sshlibrary
pymysql
reportlab>=4.0.0
python-docx>=1.0.0
openpyxl>=3.1.0
requests
